Module Arith_pvm.Arith_proof_format

This module manifests the proof format used by the Arith PVM as defined by the Layer 1 implementation for it.

It is imperative that this is aligned with the protocol's implementation.

module Tree : sig ... end

type tree = Tree.tree

type proof

val hash_tree : tree -> Tezos_crypto.Hashed.Smart_rollup_state_hash.t

val proof_encoding : proof Tezos_base.TzPervasives.Data_encoding.t

val proof_before : proof -> Tezos_crypto.Hashed.Smart_rollup_state_hash.t

val proof_after : proof -> Tezos_crypto.Hashed.Smart_rollup_state_hash.t

val produce_proof : ([ `Read | `Write ], Tezos_layer2_irmin_context__Irmin_context.repo) Tezos_layer2_store.Context_sigs.index -> tree -> (tree -> (tree * 'a) Lwt.t) -> (proof * 'a) option Lwt.t

val verify_proof : proof -> (tree -> (tree * 'a) Lwt.t) -> (tree * 'a) option Lwt.t