Module Octez_smart_rollup_node_PtParisB.Outbox

This module provides helper to interact with PVM outboxes.

val proof_of_output : Octez_smart_rollup_node.Node_context.rw -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Sc_rollup.output -> (Octez_smart_rollup.Commitment.Hash.t * string) Tezos_base.TzPervasives.tzresult Lwt.t

proof_of_output node_ctxt output returns the last cemented commitment hash and the proof of the output in the LCC.

val proof_of_output_simple : Octez_smart_rollup_node.Node_context.rw -> outbox_level:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Raw_level.t -> message_index:int -> (Octez_smart_rollup.Commitment.Hash.t * string) Tezos_base.TzPervasives.tzresult Lwt.t

proof_of_output_simple node_ctxt ~level ~message_index returns the last cemented commitment hash and the proof of the output in the LCC.